Format Changes.
Changed the Routing Cache size of GatewayTester.ino, Node0Tester.ino, Node1Tester.ino, and Node2Tester.ino from 255 to 25.

Changed the data type of PRECEIVED and PSENT of GatewayTester.ino from uint8_t to float.

Changed the number of decimals in the printing of Speed, Latitude, Longitude, PDR, and RTT of GatewayTester.ino from 10 to 6.

Added the uploading of the RTT to ThingsBoard in GatewayTester.ino.